Switched timestamp to 32-bit
32 bit is more than enough (70000 minutes-ish)

+++ src/main.c
@@ -20,7 +20,7 @@
 
 /*Init system config*/
 
-volatile uint64_t FlightTime = 0;
+volatile uint32_t FlightTime = 0;
 
 static volatile uint8_t task_100_flag = 0;
 static volatile uint8_t task_1s_flag = 0;
